refactor aux modules
